Major UI Modifications
Well this is a big commit. Moved evewything into an on page dialog box. The dialog can be disabled on a per site/global basis. Same for automatic minimisation. You can drag the ui around the page to place it wherever is most convenient (though the auto minimsation should be enough that maybe this wasn't needed. This new method allows one of the more valuable features, website end page info to load as the page loads which means it's much faster and convenient to use. A lot of options have been moved to a context sensitive section. They only appear in the app when viewing a page in the cpanel where the feature is supposed to be used.

Looks pretty ugly and more work needs to be done t ensure consistency once it is made pretty (since on page styling can now potentially affect the addon).

Oh also added a feature to automatically load the sites cPanel if one isn't alrteady open.